refactor(schema,site)!: site composition gets its own package
`routes`, `layout`, `context` and `head-merger` decide what a project's page at
a route consists of: which URL a page file has, the layout it wraps in, the
`$site`/`$page` state it binds against, and the head those merge into. None of
that is schema generation, and `@jxsuite/schema` is simply where it had
accreted — the package's own spec never described any of it.

They move to `@jxsuite/site`, a package with no platform imports at all, whose
charter is that one question. Their docblocks already argued for this home
before it existed: each says it lives outside `@jxsuite/compiler` because that
graph carries sharp and esbuild, so neither a browser bundle nor a Cloudflare
Worker can load it. Being in schema was the nearest available answer to that,
not a claim that a route is a schema.

`site-style-css` comes too, out of `packages/studio`. It is pure, it reads
`project.json`, and a second host is about to need the same stylesheet — which
takes the site-style emitters from three to two.

No re-export shim. A shim would keep the surface in schema, which is the thing
being removed.

Two guards earned their keep. `scripts/ci/bundle-paths.test.ts` caught that
studio's new dependency widened the bundle-analysis path closure, and
`check-standards.ts` caught the standards evidence in site-architecture.md
still pointing at the old paths.

BREAKING CHANGE: `@jxsuite/schema` no longer exports `./routes`, `./layout`,
`./context` or `./head-merger`. Import them from `@jxsuite/site`; the modules
and their behaviour are unchanged.
