+++ .github/workflows/bundle-analysis.yml
@@ -31,6 +31,7 @@ on:
       - "packages/protocol/**"
       - "packages/runtime/**"
       - "packages/schema/**"
+      - "packages/site/**"
       - "packages/starters/**"
       - "packages/studio/**"
       # The lockfile moves what any of them actually bundles.
