+++ docs/extending/reference/implementation-status.md
@@ -25,7 +25,7 @@ This page is generated from the `> **Status: …**` markers in the specification
 | `relationships.md`        | 0.1.4-draft  | Partial     | 2026-08-15 |
 | `schema.md`               | 0.4.8-draft  | Partial     | 2026-08-16 |
 | `server.md`               | 0.2.14       | Implemented | 2026-08-26 |
-| `site-architecture.md`    | 0.6.0-draft  | Partial     | 2026-08-26 |
+| `site-architecture.md`    | 0.6.1-draft  | Partial     | 2026-08-27 |
 | `spec.md`                 | 0.5.8-draft  | Partial     | 2026-08-26 |
 | `standards.md`            | 0.1.15-draft | Partial     | 2026-08-17 |
 | `studio-ui-guidelines.md` | 0.3.15       | Implemented | 2026-08-26 |
