+++ docs/framework/site/i18n.md
@@ -9,7 +9,7 @@ code:
   - packages/schema/src/intl.ts
   - packages/compiler/src/site/locale-negotiation.ts
   - packages/compiler/src/site/i18n.ts
-  - packages/schema/src/context.ts
+  - packages/site/src/context.ts
   - packages/compiler/src/site/context-injection.ts
 ---
