+++ packages/compiler/src/types.ts
@@ -1,8 +1,8 @@
 /**
- * `HeadMergeContext` moved to `@jxsuite/schema/head-merger` with the merger itself — a browser
- * bundle and a Worker both assemble the same `<head>`, and neither can import this package.
+ * `HeadMergeContext` moved to `@jxsuite/site/head-merger` with the merger itself — a browser bundle
+ * and a Worker both assemble the same `<head>`, and neither can import this package.
  */
-export type { HeadMergeContext } from "@jxsuite/schema/head-merger";
+export type { HeadMergeContext } from "@jxsuite/site/head-merger";
 
 export interface SiteRoute {
   urlPattern: string;
