+++ packages/schema/package.json
@@ -11,19 +11,15 @@
     ".": "./src/schema.ts",
     "./asset-paths": "./src/asset-paths.ts",
     "./component-meta": "./src/component-meta.ts",
-    "./context": "./src/context.ts",
     "./extension-registry": "./src/extension-registry.ts",
     "./format-registry": "./src/format-registry.ts",
     "./project-schemas": "./src/project-schemas.ts",
-    "./routes": "./src/routes.ts",
     "./validate-project": "./src/validate-project.ts",
     "./types": "./types.ts",
     "./guards": "./src/guards.ts",
-    "./layout": "./src/layout.ts",
     "./locale": "./src/locale.ts",
     "./intl": "./src/intl.ts",
     "./media-type": "./src/media-type.ts",
-    "./head-merger": "./src/head-merger.ts",
     "./ijson": "./src/ijson.ts",
     "./parse": "./src/parse.ts",
     "./defs": "./defs/index.ts",
