+++ packages/site/bunfig.toml
@@ -1,0 +1,8 @@
+[install]
+registry = "https://registry.npmjs.org"
+
+[test]
+coverageReporter = ["text", "lcov"]
+coverageSkipTestFiles = true
+coveragePathIgnorePatterns = ["**/tests/**", "../runtime/**", "../schema/**"]
+coverageThreshold = { lines = 0.99, functions = 0.99 }
