+++ packages/studio/package.json
@@ -65,6 +65,7 @@
     "@jxsuite/protocol": "workspace:^",
     "@jxsuite/runtime": "workspace:^",
     "@jxsuite/schema": "workspace:^",
+    "@jxsuite/site": "workspace:^",
     "@spectrum-web-components/accordion": "^1.12.2",
     "@spectrum-web-components/action-bar": "1.12.2",
     "@spectrum-web-components/action-button": "^1.12.2",
