+++ release-please-config.json
@@ -14,6 +14,10 @@
       "release-type": "node",
       "component": "schema"
     },
+    "packages/site": {
+      "release-type": "node",
+      "component": "site"
+    },
     "packages/runtime": {
       "release-type": "node",
       "component": "runtime",
